Refactor JarvisCore into new package
