+++ jarvis/__init__.py
@@ -1,6 +1,6 @@
 """JARVIS assistant package."""
 
-from .core import JarvisCore
+from jarvis_core import JarvisCore
 from .gui import JarvisGUI
 
 __all__ = ["JarvisCore", "JarvisGUI"]
