+++ jarvis_core/__init__.py
@@ -1,0 +1,5 @@
+"""Core package for the JARVIS assistant."""
+
+from .core import JarvisCore
+
+__all__ = ["JarvisCore"]
